// apps/integration-service/src/operation.types.ts
export enum OperationStatus {
  PENDING = 'PENDING',
  PROCESSING = 'PROCESSING',
  FAILED_RETRYABLE = 'FAILED_RETRYABLE',
  FAILED_BUSINESS = 'FAILED_BUSINESS',
  FINISHED = 'FINISHED',
}

export enum OperationStep {
  VALIDATE = 'VALIDATE',
  CREATE_EXTERNAL_TRANSACTION = 'CREATE_EXTERNAL_TRANSACTION',
  CONFIRM_EXTERNAL_TRANSACTION = 'CONFIRM_EXTERNAL_TRANSACTION',
  UPDATE_INTERNAL_DATABASE = 'UPDATE_INTERNAL_DATABASE',
  FINISHED = 'FINISHED',
}

export enum EventStatus {
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
}